requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
WTForms==3.0.1
Werkzeug==2.2.3
python-dotenv==1.0.1
eventlet==0.33.3
gunicorn==20.1.0
Flask-SocketIO==5.3.4
